Improved performance for mouse move
Checks only mouse move if first click is in over library.

+++ sd-editor/src/main/java/net/sevenscales/editor/api/ISurfaceHandler.java
@@ -100,7 +100,7 @@ public interface ISurfaceHandler extends OperationTransaction, HasMouseWheelHand
 	/**
 	* This should be used always when transforming root layer. Notifies automatically HTML Layer.
 	*/
-	void setTransform(int tx, int ty);
+	void setTransform(double tx, double ty);
 	void scale(double value, boolean wheel);
 	double getScaleFactor();
 	int scaleClientX(int clientX);
